moved base files to data dir
