+++ src/clean_air/data/exceptions.py
